backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
pdfplumber==0.10.3
python-multipart==0.0.6
gunicorn==21.2.0
werkzeug==3.0.0
requests==2.33.0
flask-limiter==3.5.0

# Multi-Agent LangGraph system (pinned to installed versions)
langgraph==1.1.4
langchain==1.2.14
langchain-core==1.2.28
langchain-ollama==1.1.0
langchain-groq==1.1.2
langchain-community==0.4.1
faiss-cpu==1.13.2
sentence-transformers==5.3.0
numpy==2.4.4

# V2 – Multi-Tenant SaaS
sqlalchemy==2.0.35
psycopg2-binary==2.9.11
PyJWT==2.12.1
bcrypt==5.0.0
pydantic==2.12.5

# Gemini fallback
langchain-google-genai>=2.0.0

# Testing
pytest==8.3.3
hypothesis==6.112.2
